F: Only register version for the current project
The main challenge is that currently the plugin is partially registering it self with
subprojects, but not fully doing so. It should be updated to fully register and
handle versioning everywhere or only do it for the project configured.
This change moves it to only register for the currently configured project.

I thought of making the existing behavior conditional, however
even then it didn't feel like the correct change.