feat: move extra logging to info level
Currently the plugin prints out 5 separate log messages describing refs,
configurations, versions, and describeTagFirstParent, plus a newline.

This is in addition to printing out the actual fully resolved version.
The vast majority of the time this information is not valuable, but it
gets printed even when running other gradle tasks (like compile!) in a
project that uses this plugin.

This commit moves these messages to the info log level so they are not
printed by default all the time, and instead only become visible if you
add the --info (or lower) log level.